+++ fiberplane-hooks/src/index.ts
@@ -1,3 +1,5 @@
 export * from "./useHandler";
 export * from "./useKeyPressEvent";
 export * from "./useLocalStorage";
+export * from "./useMedia";
+export * from "./useThemeSelect";